+++ NEWS.md
@@ -2,7 +2,9 @@
 
 # development version
 
-- Add Abdul-Aziz 2025 caspofungin ([doi:10.1128/aac.01435-24](https://doi.org/10.1128/aac.01435-24)) -- critically ill adults receiving extracorporeal membrane oxygenation.
+- Add Guo 2025 GLP-1 receptor agonist weight-reduction model-based meta-analysis
+  of 12 drugs ([doi:10.1016/j.obpill.2025.100162](https://doi.org/10.1016/j.obpill.2025.100162)) --
+  adults with overweight or obesity, pooled from 55 placebo-controlled trials.
 
 - Add Sager 2023 sotrovimab population PK and COVID-19 progression exposure-response models ([doi:10.1002/psp4.12958](https://doi.org/10.1002/psp4.12958)) -- non-hospitalized adults and adolescents with mild-to-moderate COVID-19 at high risk of progression, plus healthy volunteers.